Packs some of the many chunk shader parameters into Vector4s and tweaks ambient lighting and refraction water intensity.

+++ src/main/java/org/terasology/rendering/shader/ShaderParametersBase.java
@@ -62,7 +62,7 @@ public class ShaderParametersBase  implements IPropertyProvider, IShaderParamete
         }
 
         if (localPlayer != null) {
-            program.setInt("carryingTorch", localPlayer.isCarryingTorch() ? 1 : 0);
+            program.setFloat("carryingTorch", localPlayer.isCarryingTorch() ? 1.0f : 0.0f);
         }
 
         if (worldProvider != null) {
